refactor(formatter): extract MessageParser component
- Move message pattern detection and parsing logic into dedicated MessageParser class
- Extract methods: normalise_message, status_message, request_type, user_agent_message, remove_unprintable_characters
- Delegate process_message to FormattingComponents::MessageParser.new.parse
- Add require for new component in main module
