+++ lib/json_rails_logger.rb
@@ -7,6 +7,7 @@ require_relative 'json_rails_logger/constants'
 require_relative 'json_rails_logger/request_id_middleware'
 require_relative 'json_rails_logger/json_formatter'
 require_relative 'json_rails_logger/formatting_components/request_message_composer'
+require_relative 'json_rails_logger/formatting_components/message_parser'
 require_relative 'json_rails_logger/error'
 require_relative 'json_rails_logger/logger'
 require_relative 'json_rails_logger/version'
